rename StateMonitor to ProcessMonitor
